add minimap and player
