+++ client/components/Player.js
@@ -1,76 +1,113 @@
-import { generateRandomColor } from '../utils/color';
-
 const PLAYER_RADIUS = 25;
 
-export default class Player {
-  constructor(props) {
-    this.canvas = props.canvas;
-    this.position = props.position || { x: window.innerWidth / 2, y: window.innerHeight / 2 };
-    this.velocity = { dx: 0, dy: 0 };
-    this.theta = props.theta;
-    this.color = props.color || generateRandomColor();
-    this.points = 0;
+// eslint-disable-next-line
+export function drawPlayer(p, canvas, scale) {
+  const ctx = canvas.getContext('2d');
+  const { x, y } = p.position;
+  const playerSize = PLAYER_RADIUS / scale;
 
-    this.rightPressed = false;
-    this.leftPressed = false;
-    this.upPressed = false;
-    this.downPressed = false;
+  if (p.name !== '') {
+    console.log('draw ' + p.name + ': ' + p.position.x + ', ' + p.position.y);
 
-    this.drawPlayer = this.drawPlayer.bind(this);
-    this.keyDownHandler = this.keyDownHandler.bind(this);
-    this.keyUpHandler = this.keyUpHandler.bind(this);
-    window.addEventListener('keydown', this.keyDownHandler);
-    window.addEventListener('keyup', this.keyUpHandler);
-  }
-
-  drawPlayer() {
-    const ctx = this.canvas.getContext('2d');
-    const { x, y } = this.position;
+    // Proportions
+    const proportionBackCenter = 3 / 4;
+    const proportionWingOuterTop = 4 / 7;
+    const proportionWingOuterBottom = 5 / 6;
+    const proportionWingOuterDistance = 4 / 5;
+    const proportionWingTopInnerDistance = 7 / 10;
+    // Constants
+    const sinAngle = Math.sin(p.angle);
+    const cosAngle = Math.cos(p.angle);
+    const playerRadiusSinAngle = playerSize * sinAngle;
+    const playerRadiusCosAngle = playerSize * cosAngle;
+    const backCenterX = x - (playerRadiusSinAngle * proportionBackCenter); // determines location of the base of the rocket
+    const backCenterY = y - (playerRadiusCosAngle * proportionBackCenter);
+    const backLength = (playerSize / 2);
+    const backLengthSinAngle = backLength * sinAngle;
+    const backLengthCosAngle = backLength * cosAngle;
+    const wingTopX = x - (playerRadiusSinAngle / 3); // determines location of the top of the wing
+    const wingTopY = y - (playerRadiusCosAngle / 3);
+    /*
+    Start drawing Rocket Chassis, starts from bottom right to the bottom left,
+    draw toward the rocket tip then back to the bottom right to complete the shape and fill
+    */
+    // Coordinates of the Rocket Tip
+    const rocketTipX = x + (playerRadiusSinAngle * 1.2);
+    const rocketTipY = y + (playerRadiusCosAngle * 1.2);
+    // Control Points for Bezier Curve from/toward the Rocket Tip
+    const rocketTipModifierRightX = x + (playerSize * Math.sin(p.angle - Math.PI / 4));
+    const rocketTipModifierRightY = y + (playerSize * Math.cos(p.angle - Math.PI / 4));
+    const rocketTipModifierLeftX = x + (playerSize * Math.sin(p.angle + Math.PI / 4));
+    const rocketTipModifierLeftY = y + (playerSize * Math.cos(p.angle + Math.PI / 4));
+    // Center-Right Coordinates of Rocket
+    const rightCenterX = x + (playerSize * Math.sin(p.angle - Math.PI / 2));
+    const rightCenterY = y + (playerSize * Math.cos(p.angle - Math.PI / 2));
+    // Center-Left Coordinates of Rocket
+    const leftCenterX = x + (playerSize * Math.sin(p.angle + Math.PI / 2));
+    const leftCenterY = y + (playerSize * Math.cos(p.angle + Math.PI / 2));
+    // Base Coordinates
+    const rocketBottomRightX = backCenterX - backLengthCosAngle;
+    const rocketBottomRightY = backCenterY + backLengthSinAngle;
+    const rocketBottomLeftX = backCenterX + backLengthCosAngle;
+    const rocketBottomLeftY = backCenterY - backLengthSinAngle;
+    // Draw Rocket Bottom
     ctx.beginPath();
-    ctx.arc(x, y, PLAYER_RADIUS, 0, Math.PI * 2);
-    ctx.fillStyle = this.color;
+    ctx.moveTo(rocketBottomRightX, rocketBottomRightY); // bottom right side
+    ctx.lineTo(rocketBottomLeftX, rocketBottomLeftY); // bottom left side
+    // Draw Left Side
+    ctx.bezierCurveTo(leftCenterX, leftCenterY, rocketTipModifierLeftX, rocketTipModifierLeftY, rocketTipX, rocketTipY); // chassis left side
+    // Draw Right Side
+    ctx.bezierCurveTo(rocketTipModifierRightX, rocketTipModifierRightY, rightCenterX, rightCenterY, rocketBottomRightX, rocketBottomRightY); // chassis right side
+    ctx.fillStyle = p.color;
+    ctx.fill();
+    ctx.closePath();
+    /*
+    Start drawing Rocket Wings, the top of the wing is drawn first, moving toward the base of the rocket and then
+    toward the outer part of the wing before going back toward the front side and closing at the top of the wing again.
+    */
+    // Helper points along the vertical axis of the player model.
+    const wingOuterTopX = x - (playerRadiusSinAngle * proportionWingOuterTop); // Point that sets the height level of the top outer part of the wings
+    const wingOuterTopY = y - (playerRadiusCosAngle * proportionWingOuterTop);
+    const wingOuterBottomX = x - (playerRadiusSinAngle * proportionWingOuterBottom);// Point that sets the height level of the bottom outer part of the wings
+    const wingOuterBottomY = y - (playerRadiusCosAngle * proportionWingOuterBottom);
+    // Exact points for the right side of the wing
+    const wingTopRightX = wingTopX - (playerRadiusCosAngle * proportionWingTopInnerDistance); // inner top right corner
+    const wingTopRightY = wingTopY + (playerRadiusSinAngle * proportionWingTopInnerDistance);
+    const wingBotRightX = rocketBottomRightX; // inner bottom right corner
+    const wingBotRightY = rocketBottomRightY;
+    const wingOuterTopRightX = wingOuterTopX - (playerRadiusCosAngle * proportionWingOuterDistance); // outer top right corner
+    const wingOuterTopRightY = wingOuterTopY + (playerRadiusSinAngle * proportionWingOuterDistance);
+    const wingOuterBottomRightX = wingOuterBottomX - (playerRadiusCosAngle * proportionWingOuterDistance); // outer bottom right corner
+    const wingOuterBottomRightY = wingOuterBottomY + (playerRadiusSinAngle * proportionWingOuterDistance);
+    // Exact points for the left side of the wing
+    const wingTopLeftX = wingTopX + (playerRadiusCosAngle * proportionWingTopInnerDistance); // inner top left corner
+    const wingTopLeftY = wingTopY - (playerRadiusSinAngle * proportionWingTopInnerDistance);
+    const wingBotLeftX = rocketBottomLeftX; // inner bottom left corner
+    const wingBotLeftY = rocketBottomLeftY;
+    const wingOuterTopLeftX = wingOuterTopX + (playerRadiusCosAngle * proportionWingOuterDistance); // outer top left corner
+    const wingOuterTopLeftY = wingOuterTopY - (playerRadiusSinAngle * proportionWingOuterDistance);
+    const wingOuterBottomLeftX = wingOuterBottomX + (playerRadiusCosAngle * proportionWingOuterDistance); // outer bottom left corner
+    const wingOuterBottomLeftY = wingOuterBottomY - (playerRadiusSinAngle * proportionWingOuterDistance);
+    // Draw Rocket Right Wing
+    ctx.beginPath();
+    ctx.moveTo(wingTopRightX, wingTopRightY);
+    ctx.lineTo(wingBotRightX, wingBotRightY);
+    ctx.lineTo(wingOuterBottomRightX, wingOuterBottomRightY);
+    ctx.lineTo(wingOuterTopRightX, wingOuterTopRightY);
+    ctx.fillStyle = p.color;
+    ctx.fill();
+    ctx.closePath();
+    // Draw Rocket Left Wing
+    ctx.beginPath();
+    ctx.moveTo(wingTopLeftX, wingTopLeftY);
+    ctx.lineTo(wingBotLeftX, wingBotLeftY);
+    ctx.lineTo(wingOuterBottomLeftX, wingOuterBottomLeftY);
+    ctx.lineTo(wingOuterTopLeftX, wingOuterTopLeftY);
+    ctx.fillStyle = p.color;
     ctx.fill();
     ctx.closePath();
 
-    ctx.beginPath();
-    ctx.moveTo(x + (PLAYER_RADIUS * Math.sin(this.theta)), y + (PLAYER_RADIUS * Math.cos(this.theta)));
-    ctx.lineTo(x - (PLAYER_RADIUS * Math.sin(this.theta)), y - (PLAYER_RADIUS * Math.cos(this.theta)));
-    ctx.strokeStyle = '#000000';
-    ctx.strokeWidth = 5;
-    ctx.stroke();
-
-    const backCenterX = x - ((PLAYER_RADIUS * Math.sin(this.theta)) / 2);
-    const backCenterY = y - ((PLAYER_RADIUS * Math.cos(this.theta)) / 2);
-    const backLength = (2.5 * ((PLAYER_RADIUS / 2) / Math.tan(45)));
-    ctx.beginPath();
-    ctx.moveTo(backCenterX - (backLength * Math.cos(this.theta)), backCenterY + (backLength * Math.sin(this.theta)));
-    ctx.lineTo(backCenterX + (backLength * Math.cos(this.theta)), backCenterY - (backLength * Math.sin(this.theta)));
-    ctx.strokeStyle = '#0000000';
-    ctx.strokeWidth = 5;
-    ctx.stroke();
-  }
-
-  keyDownHandler(e) {
-    if (e.keyCode === 39) {
-      this.rightPressed = true;
-    } else if (e.keyCode === 37) {
-      this.leftPressed = true;
-    } else if (e.keyCode === 38) {
-      this.upPressed = true;
-    } else if (e.keyCode === 40) {
-      this.downPressed = true;
-    }
-  }
-
-  keyUpHandler(e) {
-    if (e.keyCode === 39) {
-      this.rightPressed = false;
-    } else if (e.keyCode === 37) {
-      this.leftPressed = false;
-    } else if (e.keyCode === 38) {
-      this.upPressed = false;
-    } else if (e.keyCode === 40) {
-      this.downPressed = false;
-    }
+    // TODO: Rocket Bottom piece
+    // TODO: Rocket Window
   }
 }
